Refactor/18 change package name (#19)
* refactor: 서비스명으로 패키지명 변경

- muckpot으로 변경

* refactor: cicd 스크립트 바뀐 패키지명 적용

- muckpot으로 루트 패키지명 변경 반영
