+++ muckpot-api/build.gradle.kts
@@ -1,0 +1,7 @@
+dependencies {
+    implementation(project(":muckpot-domain"))
+
+    implementation("org.springframework.boot:spring-boot-starter-web")
+    implementation("org.springframework.boot:spring-boot-starter-validation")
+    implementation("io.springfox:springfox-boot-starter:3.0.0")
+}
