+++ muckpot-api/src/main/kotlin/com/yapp/muckpot/Application.kt
@@ -1,0 +1,11 @@
+package com.yapp.muckpot
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class Application
+
+fun main(args: Array<String>) {
+    runApplication<Application>(*args)
+}
