+++ muckpot-api/src/main/kotlin/com/yapp/muckpot/swagger/ResponseExample.kt
@@ -1,0 +1,9 @@
+package com.yapp.muckpot.swagger
+
+const val TEST_SAMPLE = """
+{
+  "id": 1,
+  "name": "test",
+  "currentTime": "20230515"
+}
+"""
