+++ muckpot-api/src/test/kotlin/com/yapp/muckpot/ApplicationTests.kt
@@ -1,0 +1,12 @@
+package com.yapp.muckpot
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class ApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+}
