+++ muckpot-domain/src/main/kotlin/com/yapp/muckpot/config/QuerydslConfig.kt
@@ -1,0 +1,16 @@
+package com.yapp.muckpot.config
+
+import com.querydsl.jpa.impl.JPAQueryFactory
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import javax.persistence.EntityManager
+
+@Configuration
+class QuerydslConfig(
+    private val em: EntityManager
+) {
+    @Bean
+    fun jpaQueryFactory(): JPAQueryFactory {
+        return JPAQueryFactory(em)
+    }
+}
