+++ muckpot-domain/src/main/kotlin/com/yapp/muckpot/test/TestRepository.kt
@@ -1,0 +1,7 @@
+package com.yapp.muckpot.test
+
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface TestRepository : JpaRepository<TestEntity, Long>
