+++ settings.gradle.kts
@@ -1,7 +1,7 @@
-rootProject.name = "Web-1-BE"
+rootProject.name = "Muckpot-BE"
 
 include(
-    "web1-api",
-    "web1-domain",
-    "web1-infra"
+    "muckpot-api",
+    "muckpot-domain",
+    "muckpot-infra"
 )
