--- web1-api/build.gradle.kts
@@ -1,7 +1,0 @@
-dependencies {
-    implementation(project(":web1-domain"))
-
-    implementation("org.springframework.boot:spring-boot-starter-web")
-    implementation("org.springframework.boot:spring-boot-starter-validation")
-    implementation("io.springfox:springfox-boot-starter:3.0.0")
-}
