--- web1-api/src/main/kotlin/com/yapp/web1be/Application.kt
@@ -1,11 +1,0 @@
-package com.yapp.web1be
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class Application
-
-fun main(args: Array<String>) {
-    runApplication<Application>(*args)
-}
