--- web1-api/src/main/kotlin/com/yapp/web1be/config/SwaggerConfig.kt
@@ -1,34 +1,0 @@
-package com.yapp.web1be.config
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.bind.annotation.RestController
-import springfox.documentation.builders.ApiInfoBuilder
-import springfox.documentation.builders.PathSelectors
-import springfox.documentation.builders.RequestHandlerSelectors
-import springfox.documentation.service.ApiInfo
-import springfox.documentation.spi.DocumentationType
-import springfox.documentation.spring.web.plugins.Docket
-import springfox.documentation.swagger2.annotations.EnableSwagger2
-
-@EnableSwagger2
-@Configuration
-class SwaggerConfig {
-    @Bean
-    fun api(): Docket {
-        return Docket(DocumentationType.OAS_30)
-            .select()
-            .apis(RequestHandlerSelectors.withClassAnnotation(RestController::class.java))
-            .paths(PathSelectors.any())
-            .build()
-            .apiInfo(apiInfo())
-    }
-
-    private fun apiInfo(): ApiInfo {
-        return ApiInfoBuilder()
-            .title("먹팟")
-            .description("먹팟 API 리스트")
-            .version("0.1")
-            .build()
-    }
-}
