--- web1-api/src/main/kotlin/com/yapp/web1be/swagger/ResponseExample.kt
@@ -1,9 +1,0 @@
-package com.yapp.web1be.swagger
-
-const val TEST_SAMPLE = """
-{
-  "id": 1,
-  "name": "test",
-  "currentTime": "20230515"
-}
-"""
