--- web1-api/src/test/kotlin/com/yapp/web1be/ApplicationTests.kt
@@ -1,12 +1,0 @@
-package com.yapp.web1be
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class ApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-}
