--- web1-api/src/test/kotlin/com/yapp/web1be/test/TestServiceTest.kt
@@ -1,19 +1,0 @@
-package com.yapp.web1be.test
-
-import io.kotest.core.spec.style.StringSpec
-import io.kotest.matchers.shouldBe
-import io.mockk.every
-import io.mockk.mockk
-
-class TestServiceTest : StringSpec({
-    val testService: TestService = mockk()
-
-    "kotest 테스트 " {
-        // given
-        every { testService.test() } returns TestResponse.of(TestEntity())
-        val result = testService.test()
-
-        result.id shouldBe 1
-        result.name shouldBe "test"
-    }
-})
