--- web1-domain/src/main/kotlin/com/yapp/web1be/config/QuerydslConfig.kt
@@ -1,16 +1,0 @@
-package com.yapp.web1be.config
-
-import com.querydsl.jpa.impl.JPAQueryFactory
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import javax.persistence.EntityManager
-
-@Configuration
-class QuerydslConfig(
-    private val em: EntityManager
-) {
-    @Bean
-    fun jpaQueryFactory(): JPAQueryFactory {
-        return JPAQueryFactory(em)
-    }
-}
