--- web1-domain/src/main/kotlin/com/yapp/web1be/test/TestEntity.kt
@@ -1,11 +1,0 @@
-package com.yapp.web1be.test
-
-import javax.persistence.Entity
-import javax.persistence.Id
-
-@Entity
-data class TestEntity(
-    @Id
-    val id: Long = 1,
-    val name: String = "test"
-)
