--- web1-domain/src/main/kotlin/com/yapp/web1be/test/TestRepository.kt
@@ -1,7 +1,0 @@
-package com.yapp.web1be.test
-
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface TestRepository : JpaRepository<TestEntity, Long>
